Added calculate_solution_weights.py
This function takes in a molecular_weights dictionary whose keys are compound names and whose values are their molecular weights, and a list whose entries are in the format "compoundname-concentration" where concentration is in the format <int or float>M (e.g., "0.1M", or "1M").
The function returns a list with entries in the format "compoundname-concentration-weight" or "unknown", where weight = molecular weight * concentration. If the compoundname is not found in the molecular_weight dictionary, the entry is listed as "unknown".